Razor Hydra integration
git-svn-id: https://mediacenter-pc/svn/GPGPU/CUDA/RaytracingEngine/trunk@921 c4de0f11-4f5c-764a-a4aa-d9e147179871

+++ CPUKernel.cpp
@@ -379,8 +379,8 @@ float4 CPUKernel::cubeMapping(
 #ifdef USE_KINECT
    if( primitive.type.x == ptCamera )
    {
-      int x = (intersection.x-primitive.p0.x+primitive.size.x)*material.textureMapping.x;
-      int y = gKinectVideoHeight - (intersection.y-primitive.p0.y+primitive.size.y)*material.textureMapping.y;
+      int x = static_cast<int>((intersection.x-primitive.p0.x+primitive.size.x)*material.textureMapping.x);
+      int y = static_cast<int>(gKinectVideoHeight - (intersection.y-primitive.p0.y+primitive.size.y)*material.textureMapping.y);
 
       x = (x+gKinectVideoWidth)%gKinectVideoWidth;
       y = (y+gKinectVideoHeight)%gKinectVideoHeight;
